+++ config/robots.js
@@ -1,0 +1,11 @@
+/**
+ * @file define robots.
+ */
+
+module.exports = {
+
+  useragent: '*',
+  disallow: ['/apple-app-site-association', '/.well-known/'],
+  sitemap: 'https://example.com/sitemap.xml.gz',
+
+};